+++ extensions/supermemory/CHANGELOG.md
@@ -1,9 +1,11 @@
 # Supermemory Changelog
 
-## [] - {PR_MERGE_DATE}
+## [Search Projects + Enhancements] - {PR_MERGE_DATE}
 
+- Added command to search and add projects
 - Removed `useEffect`
 - Simplified `getApiKey` since `Preferences` will be enforced for presence and trim automatically
+- Moved API Key check into its own HoC
 
 ## [Initial Version] - 2025-10-02